% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Run the Inputs Selection App}
\usage{
run_app(...)
}
\arguments{
\item{...}{A list of options to pass to the shiny app. See \code{?shiny::shinyApp()} for details.}
}
\description{
Runs the inputs selection app in a shiny session. This function is called by
the \code{run_app()} function in the \code{app.R} file.
}
