% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upgrade_params.R
\name{upgrade_params}
\alias{upgrade_params}
\title{Upgrade model parameters to the latest version}
\usage{
upgrade_params(p)
}
\arguments{
\item{p}{A list of model parameters to upgrade.}
}
\description{
Older model parameters are automatically upgraded to the latest version when loaded. This function is called by
\code{load_params()} in \code{ZZZ.R} and should not be called directly.
}
